requirements.txt
fastmcp>=0.2.0
requests>=2.31.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psutil>=5.9.0
